wasm: Use f32 coordinates throughout interaction layer #BDK-338
- change TouchEvent, Rect, and InteractionState from i32 to f32
- eliminate lossy f64→i32 truncation at the wl_touch→runtime boundary
- eliminate redundant i32→f32 casts in button_with_pos and get_drag_pos
- remove cast_possible_wrap workaround from Rect::contains
- update testbed mouse_pos and hit_test_tile to f32
- scroll delta is now f32, removing integer quantization artifacts

+++ bmc-wasm-runtime/src/components/button.rs
@@ -3,7 +3,6 @@
 //! Button component with immediate-mode API.
 
 #![allow(clippy::wildcard_imports)]
-#![expect(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
 
 use crate::colors::*;
 use crate::interaction::{InteractionState, Rect};
@@ -187,7 +186,7 @@ pub fn draw_button(
     disabled: bool,
     skin: Option<&ButtonSkinData>,
 ) -> (bool, Option<(f32, f32)>) {
-    let bounds = Rect::new(x as i32, y as i32, w as u32, h as u32);
+    let bounds = Rect::new(x, y, w, h);
 
     if disabled {
         renderer.push_scissor(x, y, w, h);
